fix some crashing

+++ index.ts
@@ -200,35 +200,39 @@ client.on("guildCreate", (guild) => {
 
 //actions to run when the bot leaves a server
 client.on("guildDelete", async (guild) => {
-  let d = new Date();
-  const Embed = new MessageEmbed()
-    .setColor("#ff3505")
-    .setTitle(`I Left a Server`)
-    .setThumbnail(
-      String(
-        guild.iconURL() || "https://cdn.discordapp.com/embed/avatars/0.png"
+  if(guild.available) {
+    const doc = await guildModel.findOne({ GUILD_ID: guild.id });
+    console.log(guild)
+    const d = new Date();
+    const Embed = new MessageEmbed()
+      .setColor("#ff3505")
+      .setTitle(`I Left a Server`)
+      .setThumbnail(
+        String(
+          guild.iconURL() || "https://cdn.discordapp.com/embed/avatars/0.png"
+        )
       )
-    )
-    .addFields([
-      { name: "Guild Creation Date:", value: guild.createdAt.toString() },
-      { name: "Guild Leave Date:", value: d.toString() },
-      { name: "Guild Name:", value: guild.name },
-      { name: "Guild ID:", value: guild.id },
-      { name: "Owner ID:", value: guild.ownerId },
-      { name: "Guild Member Count:", value: String(guild.memberCount) },
-    ])
-    .setFooter(
-      `Delivered in: ${client.ws.ping}ms | Antares Bot | ${process.env.VERSION}`,
-      "https://playantares.com/resources/icon.png"
-    );
-  try {
-    await guildModel.findOneAndDelete({ GUILD_ID: guild.id });
-    await piiModel.findOneAndDelete({ GUILD_ID: guild.id });
-    client.users.fetch(String(process.env.BOT_OWNER_ID)).then((user) => {
-      user.send({ embeds: [Embed] });
-    });
-  } catch (e) {
-    console.log(e);
+      .addFields([
+        { name: "Guild Creation Date:", value: guild.createdAt.toString() },
+        { name: "Guild Leave Date:", value: d.toString() },
+        { name: "Guild Name:", value: guild.name },
+        { name: "Guild ID:", value: guild.id },
+        { name: "Owner ID:", value: guild.ownerId },
+        { name: "Guild Member Count:", value: doc?.GUILD_MEMBER_COUNT },
+      ])
+      .setFooter(
+        `Delivered in: ${client.ws.ping}ms | Antares Bot | ${process.env.VERSION}`,
+        "https://playantares.com/resources/icon.png"
+      );
+    try {
+      await guildModel.findOneAndDelete({ GUILD_ID: guild.id });
+      await piiModel.findOneAndDelete({ GUILD_ID: guild.id });
+      client.users.fetch(String(process.env.BOT_OWNER_ID)).then((user) => {
+        user.send({ embeds: [Embed] });
+      });
+    } catch (e) {
+      console.log(e);
+    }
   }
 });
